+++ lib/src/code_generators/constants.dart
@@ -20,6 +20,7 @@ const kChopperApi = 'ChopperApi';
 const kResponse = 'Response';
 const kFutureResponse = 'Future<chopper.Response>';
 const kString = 'string';
+const kUndefinedParameter = 'undefinedParameter';
 
 const kResponses = '/responses/';